# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIREST Sistema de Pacientes",
        "description": "Documentación de la API para gestionar el sistema pacientes",
        "contact": {
            "name": "Gustavo Galarza- Daniel Mamani",
            "email": "https://github.com/GustavoGalarza/sistema-pos.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/citas": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Obtener lista de citas",
                "description": "Devuelve una lista de citas paginadas (Requiere autenticación)",
                "operationId": "bc2ee81fe758951b2eaebf55a4ed4613",
                "responses": {
                    "200": {
                        "description": "Lista de citas obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "paciente_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "medico_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "fecha_cita": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-20T10:00:00"
                                                    },
                                                    "motivo": {
                                                        "type": "string",
                                                        "example": "Consulta general"
                                                    },
                                                    "pacientes": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "medicos": {
                                                        "type": "string",
                                                        "example": "Dr. López"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Citas"
                ],
                "summary": "Crear una nueva cita",
                "description": "Registra una nueva cita (Requiere autenticación)",
                "operationId": "a82bd38c189f57b1e8e24bd3154053cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paciente_id",
                                    "medico_id",
                                    "fecha_cita",
                                    "motivo"
                                ],
                                "properties": {
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_cita": {
                                        "type": "string",
                                        "example": "2024-10-20 10:00:00"
                                    },
                                    "motivo": {
                                        "type": "string",
                                        "example": "Consulta general"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cita creada satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cita creado satisfactoriamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/citas/pacientes/{paciente_id}": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Obtener citas por paciente",
                "description": "Devuelve todas las citas de un paciente específico (Requiere autenticación)",
                "operationId": "5807596f2b799df13a9e48ac307dd7e8",
                "parameters": [
                    {
                        "name": "paciente_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Citas encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paciente no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}